+++ src/common/styles/ukraine-page.css
@@ -15,3 +15,12 @@
     margin: 0;
     padding: 0;
 }
+
+.table-of-contents-wrapper {
+    position: relative;
+}
+
+.table-of-contents {
+    position: sticky;
+    top: 0;
+}
